updates to rolling file name

+++ R - hei_aim3a/r0_temporal_adjustment.R
@@ -100,35 +100,39 @@ message("loading 1s onroad data")
 add_progress_notes("loading 1s onroad data")
 
 if(!file.exists(file.path(dt_pt2, "TEMP_road_dt.rda")) | !file.exists(file.path(dt_pt2, "TEMP_road_dt_no_hwy.rda"))) {
-  road_dt0 <- readRDS(file.path("data", "onroad", "annie", "OnRoad Paper Code Data", "data", "All_Onroad_12.20.rds")) %>%
-    select(-c(native_id, contains("gps_"), LABEL, num_visit, median_measurement,exclude_flag,update_exclude_flag, speed, ufp_median_5min, A1_flag, hour, minute, second)) %>%
-    arrange(time)  
-
-  missingness_table <- count_missingness(road_dt0, notes="all data")
   
-  segment_lat_long <- distinct(road_dt0, id, mx, my, road_type)
-  saveRDS(segment_lat_long, file.path(dt_pt, "segment_lat_long.rda"))
-  rm(segment_lat_long)
+  if(!file.exists(file.path(dt_pt2, "TEMP_road_dt0.rda"))) {
+    road_dt0 <- readRDS(file.path("data", "onroad", "annie", "OnRoad Paper Code Data", "data", "All_Onroad_12.20.rds")) %>%
+      select(-c(native_id, contains("gps_"), LABEL, num_visit, median_measurement,exclude_flag,update_exclude_flag, speed, ufp_median_5min, A1_flag, hour, minute, second)) %>%
+      arrange(time)  
   
-  road_dt0 <- select(road_dt0, -c(mx,my))
+    missingness_table <- count_missingness(road_dt0, notes="all data")
     
-  road_dt0 <- road_dt0 %>%
-    drop_na(ufp)
-  missingness_table <- count_missingness(road_dt0, notes="missing 'ufp'")
-  
-  # some duplicate rows. some remain if same value was assigned to 2 locations during the same second
-  road_dt0 <- distinct(road_dt0)
-  missingness_table <- count_missingness(road_dt0, notes="drop duplicate rows")
-  
-  #only keep the first segment/reading when the same reading is assigned to 2 locations/neighboring segments?
-  road_dt0 <- road_dt0 %>%
-    group_by(time) %>%
-    slice(1) %>%
-    ungroup()  
-  
-  saveRDS(road_dt0, file.path(dt_pt2, "TEMP_road_dt0.rda"))
-  
-  missingness_table <- count_missingness(road_dt0, notes="only keep 1st location when same sec reading is assigned to 2")
+    segment_lat_long <- distinct(road_dt0, id, mx, my, road_type)
+    saveRDS(segment_lat_long, file.path(dt_pt, "segment_lat_long.rda"))
+    rm(segment_lat_long)
+    
+    road_dt0 <- select(road_dt0, -c(mx,my))
+      
+    road_dt0 <- road_dt0 %>%
+      drop_na(ufp)
+    missingness_table <- count_missingness(road_dt0, notes="missing 'ufp'")
+    
+    # some duplicate rows. some remain if same value was assigned to 2 locations during the same second
+    road_dt0 <- distinct(road_dt0)
+    missingness_table <- count_missingness(road_dt0, notes="drop duplicate rows")
+    
+    #only keep the first segment/reading when the same reading is assigned to 2 locations/neighboring segments?
+    road_dt0 <- road_dt0 %>%
+      group_by(time) %>%
+      slice(1) %>%
+      ungroup()  
+    missingness_table <- count_missingness(road_dt0, notes="only keep 1st location when same sec reading is assigned to 2")
+    write.csv(missingness_table, file.path(dt_pt2, "missing_counts_second_dt_cleanup.csv"), row.names = F)
+    
+    saveRDS(road_dt0, file.path(dt_pt2, "TEMP_road_dt0.rda"))
+    
+    } else {road_dt0 <- readRDS(file.path(dt_pt2, "TEMP_road_dt0.rda"))}
   
   # drop roosevelt garage
   road_dt0 <- road_dt0 %>%
@@ -141,6 +145,9 @@ if(!file.exists(file.path(dt_pt2, "TEMP_road_dt.rda")) | !file.exists(file.path(
     mutate(ufp = winsorize(ufp, minval=quantile(ufp, 0.05, na.rm=T), maxval=quantile(ufp, 0.95, na.rm=T))) %>%
     ungroup()
     
+  saveRDS(road_dt0, file.path(dt_pt2, "TEMP_road_dt0_winsorized.rda"))
+  # road_dt0 <- readRDS(file.path(dt_pt2, "TEMP_road_dt0_winsorized.rda"))
+  
   # drop hwy for some of these
   road_dt0_no_hwy <- filter(road_dt0, road_type != "A1") %>%
     select(-road_type)
@@ -148,7 +155,6 @@ if(!file.exists(file.path(dt_pt2, "TEMP_road_dt.rda")) | !file.exists(file.path(
   road_dt0 <- select(road_dt0, -road_type) 
   
   missingness_table <- count_missingness(road_dt0_no_hwy, notes="drop A1 segments")
-  
   write.csv(missingness_table, file.path(dt_pt2, "missing_counts_second_dt.csv"), row.names = F)
   
   # x=unique(road_dt0$runname)[1]
@@ -223,18 +229,19 @@ quantiles <- c(0.01, 0.05)
  #dt=road_dt_no_hwy %>% filter(runname == first(runname))
 # windows.=windows 
 # quantiles.=quantiles 
-calculate_rolling_quantile <- function(dt, windows.=windows, quantiles.=quantiles) {
+# file_label="_no_hwy"
+# w=windows.[1]
+# p=quantiles.[1]
+calculate_rolling_quantile <- function(dt, windows.=windows, quantiles.=quantiles, file_label="") {
   
   adj_lvls <- apply(expand.grid(paste0("hr", sort(windows)/3600), 
                                 paste0("_pct", sort(quantiles)*100)), 
                     1, paste, collapse="")
   
-  # w=windows.[1]
-  # p=quantiles.[1]
   dt <- lapply(windows., function(w) {
     mclapply(quantiles., mc.cores=use_cores, function(p) {
       
-      bg_label <- paste0("hr", w/3600, "_pct", p*100)
+      bg_label <- paste0("hr", w/3600, "_pct", p*100, file_label)
       print(paste(Sys.time(), bg_label))
       
       file_name <- file.path(dt_pt2, paste0("uw_temp_adj_1s_", bg_label, ".rda"))
@@ -264,7 +271,7 @@ saveRDS(road_dt, file.path(dt_pt2, "underwrite_temp_adj_all_1s_data.rda"))
 
 message("running underwrite temporal adjustment for non-highway data")
 add_progress_notes("running underwrite tmeporal adjustment for non-highway data")
-road_dt_no_hwy <- calculate_rolling_quantile(dt=road_dt_no_hwy)
+road_dt_no_hwy <- calculate_rolling_quantile(dt=road_dt_no_hwy, file_label="_no_hwy")
 saveRDS(road_dt_no_hwy, file.path(dt_pt2, "underwrite_temp_adj_all_1s_data_no_hwy.rda")) 
 
 ##################################################################################################
